Adjust settings view and translations
