+++ app/src/main/AndroidManifest.xml
@@ -347,7 +347,7 @@
         <activity
             android:name="com.best.deskclock.settings.DigitalWidgetCustomizationActivity"
             android:excludeFromRecents="true"
-            android:label="@string/settings_digital_widget_title"
+            android:label="@string/digital_widget_title"
             android:exported="true"
             android:theme="@style/Theme.DeskClock" />
 
@@ -386,7 +386,7 @@
         <activity
             android:name="com.best.deskclock.settings.DigitalWidgetMaterialYouCustomizationActivity"
             android:excludeFromRecents="true"
-            android:label="@string/settings_digital_widget_material_you_title"
+            android:label="@string/digital_widget_material_you_title"
             android:exported="true"
             android:theme="@style/Theme.DeskClock" />
