+++ app/src/main/java/com/best/deskclock/settings/AlarmSettingsActivity.java
@@ -68,6 +68,9 @@ public class AlarmSettingsActivity extends CollapsingToolbarBaseActivity {
         @Override
         public void onResume() {
             super.onResume();
+            int bottomPadding = Utils.toPixel(20, requireContext());
+            getListView().setPadding(0, 0, 0, bottomPadding);
+
             refresh();
         }
