+++ app/src/main/java/com/best/deskclock/settings/ClockSettingsActivity.java
@@ -61,6 +61,9 @@ public class ClockSettingsActivity extends CollapsingToolbarBaseActivity {
         @Override
         public void onResume() {
             super.onResume();
+            int bottomPadding = Utils.toPixel(20, requireContext());
+            getListView().setPadding(0, 0, 0, bottomPadding);
+
             refresh();
         }
