+++ app/src/main/java/com/best/deskclock/settings/DigitalWidgetCustomizationActivity.java
@@ -133,6 +133,9 @@ public class DigitalWidgetCustomizationActivity extends CollapsingToolbarBaseAct
         @Override
         public void onResume() {
             super.onResume();
+            int bottomPadding = Utils.toPixel(20, requireContext());
+            getListView().setPadding(0, 0, 0, bottomPadding);
+
             refresh();
             updateDigitalWidget();
         }
@@ -176,11 +179,11 @@ public class DigitalWidgetCustomizationActivity extends CollapsingToolbarBaseAct
                         mDigitalWidgetMaxClockFontSizePref.setEnabled(areCitiesDisplayed);
                         if (!areCitiesDisplayed) {
                             mDigitalWidgetMaxClockFontSizePref.setSummary(
-                                    requireContext().getString(R.string.settings_digital_widget_message_summary)
+                                    requireContext().getString(R.string.digital_widget_message_summary)
                             );
                         } else {
                             mDigitalWidgetMaxClockFontSizePref.setSummary(
-                                    requireContext().getString(R.string.settings_digital_widget_max_clock_font_size_summary)
+                                    requireContext().getString(R.string.digital_widget_max_clock_font_size_summary)
                                             + DataModel.getDataModel().getDigitalWidgetMaxClockFontSize()
                             );
                         }
@@ -254,7 +257,7 @@ public class DigitalWidgetCustomizationActivity extends CollapsingToolbarBaseAct
                 case KEY_DIGITAL_WIDGET_MAX_CLOCK_FONT_SIZE -> {
                     final EditTextPreference digitalWidgetMaxClockFontSizePref = (EditTextPreference) pref;
                     digitalWidgetMaxClockFontSizePref.setSummary(
-                            requireContext().getString(R.string.settings_digital_widget_max_clock_font_size_summary)
+                            requireContext().getString(R.string.digital_widget_max_clock_font_size_summary)
                                     + newValue.toString()
                     );
                     requireContext().sendBroadcast(new Intent(ACTION_DIGITAL_WIDGET_CLOCK_FONT_SIZE_CHANGED));
@@ -318,19 +321,19 @@ public class DigitalWidgetCustomizationActivity extends CollapsingToolbarBaseAct
                 if (mShowCitiesOnDigitalWidgetPref.isChecked()) {
                     mDigitalWidgetMaxClockFontSizePref.setEnabled(false);
                     mDigitalWidgetMaxClockFontSizePref.setSummary(
-                            requireContext().getString(R.string.settings_digital_widget_message_summary)
+                            requireContext().getString(R.string.digital_widget_message_summary)
                     );
                 } else {
                     mDigitalWidgetMaxClockFontSizePref.setEnabled(true);
                     mDigitalWidgetMaxClockFontSizePref.setSummary(
-                            requireContext().getString(R.string.settings_digital_widget_max_clock_font_size_summary)
+                            requireContext().getString(R.string.digital_widget_max_clock_font_size_summary)
                                     + DataModel.getDataModel().getDigitalWidgetMaxClockFontSize()
                     );
                 }
             } else {
                 mDigitalWidgetMaxClockFontSizePref.setEnabled(true);
                 mDigitalWidgetMaxClockFontSizePref.setSummary(
-                        requireContext().getString(R.string.settings_digital_widget_max_clock_font_size_summary)
+                        requireContext().getString(R.string.digital_widget_max_clock_font_size_summary)
                                 + DataModel.getDataModel().getDigitalWidgetMaxClockFontSize()
                 );
             }
