+++ app/src/main/java/com/best/deskclock/settings/InterfaceCustomizationActivity.java
@@ -75,6 +75,9 @@ public class InterfaceCustomizationActivity extends CollapsingToolbarBaseActivit
         @Override
         public void onResume() {
             super.onResume();
+            int bottomPadding = Utils.toPixel(20, requireContext());
+            getListView().setPadding(0, 0, 0, bottomPadding);
+
             refresh();
         }
