+++ app/src/main/java/com/best/deskclock/settings/TimerSettingsActivity.java
@@ -58,6 +58,9 @@ public class TimerSettingsActivity extends CollapsingToolbarBaseActivity {
         @Override
         public void onResume() {
             super.onResume();
+            int bottomPadding = Utils.toPixel(20, requireContext());
+            getListView().setPadding(0, 0, 0, bottomPadding);
+
             refresh();
         }
