+++ app/src/main/java/com/best/deskclock/widget/CollapsingToolbarBaseActivity.java
@@ -8,8 +8,11 @@ package com.best.deskclock.widget;
 
 import static com.best.deskclock.settings.InterfaceCustomizationActivity.KEY_AMOLED_DARK_MODE;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 
@@ -22,6 +25,9 @@ import androidx.coordinatorlayout.widget.CoordinatorLayout;
 import com.best.deskclock.R;
 import com.best.deskclock.Utils;
 import com.best.deskclock.data.DataModel;
+import com.best.deskclock.ringtone.RingtonePickerActivity;
+import com.best.deskclock.settings.AboutActivity;
+import com.best.deskclock.worldclock.CitySelectionActivity;
 import com.google.android.material.appbar.AppBarLayout;
 import com.google.android.material.appbar.CollapsingToolbarLayout;
 
@@ -59,6 +65,30 @@ public class CollapsingToolbarBaseActivity extends AppCompatActivity {
 
         final Toolbar toolbar = findViewById(R.id.action_bar);
         setSupportActionBar(toolbar);
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        if (!(this instanceof RingtonePickerActivity || this instanceof CitySelectionActivity
+                || this instanceof AboutActivity)) {
+
+            menu.add(0, Menu.NONE, 0, R.string.about_title)
+                    .setIcon(R.drawable.ic_about).setShowAsAction(MenuItem.SHOW_AS_ACTION_ALWAYS);
+            return true;
+        }
+
+        return false;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if (item.getItemId() == 0) {
+            final Intent settingIntent = new Intent(getApplicationContext(), AboutActivity.class);
+            startActivity(settingIntent);
+            return true;
+        }
+
+        return super.onOptionsItemSelected(item);
     }
 
     @Override
