+++ app/src/main/res/layout/permissions_management_activity.xml
@@ -14,7 +14,8 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginHorizontal="10dp"
-        android:layout_marginTop="20dp"
+        android:layout_marginTop="25dp"
+        android:layout_marginBottom="15dp"
         android:orientation="vertical">
 
         <com.google.android.material.card.MaterialCardView
